Feat: Add release.sh script for automated app and DMG building
This commit introduces a  script to automate the process of building the macOS application bundle () using PyInstaller and then packaging it into a Disk Image () using Not enough arguments. Run 'create-dmg --help' for help..

The script streamlines the release process by:
- Cleaning up previous build artifacts ( and  directories).
- Running PyInstaller with the configured  file to create the  bundle.
- Utilizing Not enough arguments. Run 'create-dmg --help' for help. to generate a lightweight  installer.

This automation ensures a consistent and efficient packaging workflow for the application.